pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsdom-api"
version = "0.2.0"
description = "DOM-style parser API for scanned Japanese newspaper PDFs"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = "MIT"
authors = [{name = "Seongho Bae"}]
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "pydantic>=2.9,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "reportlab>=4.2,<5.0",
  "Pillow>=11.0,<13.0",
  "pypdf>=6.10.2,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<10.0",
  "pytest-cov>=5.0,<8.0",
  "pytest-asyncio>=0.21",
  "httpx>=0.28,<1.0",
  "pyyaml>=6.0,<7.0",
]
mineru = [
  "mineru[pipeline]==3.0.9",
  "transformers>=4.48.0",
]
docs = [
    "mkdocs>=1.6,<2.0",
    "mkdocs-material>=9.6,<9.7",
]
fuzz = [
  "atheris==3.0.0 ; platform_system == 'Linux' and python_version >= '3.11'",
  "pyinstaller==6.16.0",
]

[project.scripts]
mineru = "mineru.cli.app:app"

[tool.setuptools]
package-dir = { "" = "src", "tools" = "tools"}

[tool.setuptools.packages.find]
where = ["src", "tools"]
include = ["newsdom_api*", "tools*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore:the load_module\\(\\) method is deprecated:DeprecationWarning"
]
markers = [
  "integration: requires MinerU CLI and model cache"
]

[tool.coverage.run]
source = ["src/newsdom_api", "tools"]
branch = true

[tool.coverage.report]
fail_under = 100
show_missing = true
skip_covered = false

[tool.mypy]
python_version = "3.10"
exclude = "(^tests/|^fuzzers/|^research_repos/)"

[[tool.mypy.overrides]]
module = [
  "fastapi",
  "fastapi.*",
  "pydantic",
  "pydantic.*",
  "PIL",
  "PIL.*",
  "reportlab",
  "reportlab.*",
]
ignore_missing_imports = true
